Uses Vault helpers
